move monitor into subdir
