+++ app/src/monitor/MonitorDeleteComponent.tsx
@@ -2,7 +2,7 @@ import { Navigate, useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import Axios from 'axios';
 
-import { Loading, RouteTitle } from './components';
+import { Loading, RouteTitle } from '../components';
 
 function useDelete(path: string) {
   const [done, setDone] = useState(false);
